refac(crypto): extract `PoseidonState` from `poseidon.h`
