+++ tachyon/crypto/hashes/sponge/poseidon/BUILD.bazel
@@ -7,6 +7,7 @@ tachyon_cc_library(
     hdrs = ["poseidon.h"],
     deps = [
         ":poseidon_config",
+        ":poseidon_state",
         "//tachyon/base/buffer:copyable",
         "//tachyon/crypto/hashes:prime_field_serializable",
         "//tachyon/crypto/hashes/sponge",
@@ -52,6 +53,15 @@ tachyon_cc_library(
         "//tachyon/math/base:big_int",
         "//tachyon/math/matrix:matrix_types",
         "//tachyon/math/matrix:prime_field_num_traits",
+    ],
+)
+
+tachyon_cc_library(
+    name = "poseidon_state",
+    hdrs = ["poseidon_state.h"],
+    deps = [
+        "//tachyon/crypto/hashes/sponge",
+        "//tachyon/math/matrix:matrix_types",
     ],
 )
